extract to new file
